refactor(codegen-new): split stmt.rs (848) into stmt_let + stmt_assign (<500 each)
stmt.rs had grown past the 500-line module rule (phases 2+4 pushed it to 848).
Carve the cohesive local-mutation clusters into sibling modules:
- stmt_let.rs   (313) — lower_let + clear_local_classifications + bind_tagged_local
- stmt_assign.rs(230) — lower_assign + the local `++`/`--` lowering
stmt.rs (341) keeps the stmt dispatcher, lower_return, member/index inc-dec, and
control flow (if/while). lower_let / lower_member_index_incdec became pub(super)
for the cross-module calls; the free helper ident_target is imported from stmt.

No behavior change (pure code move + import trim). Suite: rts-codegen-new lib
794 passed / 0 failed (2 ignored). Gate: stmt.rs off the >500 list (list shrank).
First of the Phase-5 file splits.
